+++ inc/template-functions.php
@@ -4,7 +4,13 @@
  * Functions which enhance the theme by hooking into WordPress
  *
  * @package Bootscore
+ * @version 5.3.3
  */
+
+
+// Exit if accessed directly
+defined( 'ABSPATH' ) || exit;
+
 
 /**
  * Adds custom classes to the array of body classes.
@@ -40,20 +46,3 @@ function bootscore_pingback_header() {
 }
 
 add_action('wp_head', 'bootscore_pingback_header');
-
-
-/**
- * Make main content col dynamic if sidebar widgets exists
- * @return string
- */
-if (!function_exists('bootscore_main_col_class')) {
-  function bootscore_main_col_class() {
-    if (!is_active_sidebar('sidebar-1')) {
-      // Sidebar is empty
-      return "col";
-    } else {
-      // Sidebar has widgets
-      return "col-md-8 col-lg-9";
-    }
-  }
-}
